requirements.txt
opencv-contrib-python==4.1.0.25
opencv-python==4.1.0.25
Pillow==7.1.1
numpy==1.18.3